src/requirements.txt
openai
pymongo
uvicorn
fastapi
python-dotenv